refactor: move DownloadButton component to Button folder
